refactor(playground): split CanvasRenderer into render/ layers
Replace the 1192-line canvas.ts monolith with src/render/:

render/renderer.ts — CanvasRenderer orchestrator (497 lines)
render/layout.ts — Scale, scaleFor, findNearestStop
render/palette.ts — all color/sizing constants
render/color-utils.ts — shade, hexWithAlpha, withAlpha, arcPoint, easeOutNorm
render/primitives.ts — roundedRect, truncate
render/draw-building.ts — shaft channels/labels, floors, gutters, waiting figs
render/draw-cars.ts — target markers, car trail, car body, bubbles
render/figures/ — stick-figure, figure-row, riders-in-car

Zero rendering behavior changes. Bundle size unchanged (86.33 KB).
