+++ playground/src/render/index.ts
@@ -1,0 +1,1 @@
+export { CanvasRenderer } from "./renderer";
