+++ playground/src/render/primitives.ts
@@ -1,0 +1,47 @@
+/**
+ * Build a rounded-rectangle path on `ctx`. Caller fills/strokes.
+ * Uses `CanvasRenderingContext2D.roundRect` when available
+ * (Chrome/Edge 99+, Safari 16+, Firefox 113+) and falls back to a
+ * manual path for older engines. The playground's Vite + esbuild
+ * target is modern browsers; the fallback keeps a local dev build on
+ * an older headless Chromium (e.g. CI screenshotters) working.
+ */
+export function roundedRect(
+  ctx: CanvasRenderingContext2D,
+  x: number,
+  y: number,
+  w: number,
+  h: number,
+  r: number,
+): void {
+  const rr = Math.min(r, w / 2, h / 2);
+  ctx.beginPath();
+  if (typeof ctx.roundRect === "function") {
+    ctx.roundRect(x, y, w, h, rr);
+    return;
+  }
+  ctx.moveTo(x + rr, y);
+  ctx.lineTo(x + w - rr, y);
+  ctx.quadraticCurveTo(x + w, y, x + w, y + rr);
+  ctx.lineTo(x + w, y + h - rr);
+  ctx.quadraticCurveTo(x + w, y + h, x + w - rr, y + h);
+  ctx.lineTo(x + rr, y + h);
+  ctx.quadraticCurveTo(x, y + h, x, y + h - rr);
+  ctx.lineTo(x, y + rr);
+  ctx.quadraticCurveTo(x, y, x + rr, y);
+  ctx.closePath();
+}
+
+/** Truncate `text` to fit `maxWidth` px, appending "\u2026" if needed. */
+export function truncate(ctx: CanvasRenderingContext2D, text: string, maxWidth: number): string {
+  if (ctx.measureText(text).width <= maxWidth) return text;
+  const ellipsis = "\u2026";
+  let lo = 0;
+  let hi = text.length;
+  while (lo < hi) {
+    const mid = (lo + hi + 1) >> 1;
+    if (ctx.measureText(text.slice(0, mid) + ellipsis).width <= maxWidth) lo = mid;
+    else hi = mid - 1;
+  }
+  return lo === 0 ? ellipsis : text.slice(0, lo) + ellipsis;
+}
